+++ kwimage/structs/heatmap.py
@@ -103,7 +103,7 @@ class _HeatmapDrawMixin(object):
     def _colorize_class_idx(self):
         """
         """
-        cidxs = kwarray.ArrayAPI.numpy(self.data['class_idx']).astype(np.int)
+        cidxs = kwarray.ArrayAPI.numpy(self.data['class_idx']).astype(int)
 
         import networkx as nx
         import kwimage
@@ -621,9 +621,9 @@ class _HeatmapWarpMixin(object):
         import kwimage
         M = self.tf_data_to_img.params[0:3]
         dsize = tuple(map(int, self.img_dims[::-1]))
-        # flags = kwimage.im_cv2._rectify_interpolation('lanczos')
-        # flags = kwimage.im_cv2._rectify_interpolation('nearest')
-        flags = kwimage.im_cv2._rectify_interpolation(interpolation)
+        # flags = kwimage.im_cv2._coerce_interpolation('lanczos')
+        # flags = kwimage.im_cv2._coerce_interpolation('nearest')
+        flags = kwimage.im_cv2._coerce_interpolation(interpolation)
         aligned = cv2.warpAffine(mask, M[0:2], dsize=tuple(dsize), flags=flags)
         aligned = np.clip(aligned, 0, 1)
         return aligned
@@ -720,6 +720,7 @@ class _HeatmapWarpMixin(object):
             >>> # xdoctest: +REQUIRES(module:ndsampler)
             >>> self = kwimage.Heatmap.random(dims=(100, 100), dets='coco', keypoints=True)
             >>> image = np.zeros(self.img_dims)
+            >>> # xdoctest: +REQUIRES(module:kwplot)
             >>> toshow = self.draw_on(image, 1, vecs=True, with_alpha=0.85)
             >>> # xdoctest: +REQUIRES(--show)
             >>> import kwplot
@@ -779,7 +780,7 @@ class _HeatmapWarpMixin(object):
                 ]))
                 corners2 = corners.warp(mat.numpy())
                 wh2 = corners2.data.clip(1, None).max(axis=0)
-                w2, h2 = np.ceil(wh2).astype(np.int).tolist()
+                w2, h2 = np.ceil(wh2).astype(int).tolist()
                 output_dims = (w2, h2)
                 return output_dims
             output_dims = _auto_select_warped_output_shape(mat_notrans)
@@ -1297,6 +1298,7 @@ class Heatmap(_generic.Spatial, _HeatmapDrawMixin,
             >>> self = kwimage.Heatmap.random(dims=(50, 200), dets='coco',
             >>>                               keypoints=True)
             >>> image = np.zeros(self.img_dims)
+            >>> # xdoctest: +REQUIRES(module:kwplot)
             >>> toshow = self.draw_on(image, 1, vecs=True, kpts=0, with_alpha=0.85)
             >>> # xdoctest: +REQUIRES(--show)
             >>> import kwplot
@@ -1340,7 +1342,7 @@ class Heatmap(_generic.Spatial, _HeatmapDrawMixin,
                 scale=scale, translation=translation)
 
             wh_dims = dims[::-1]
-            img_wh_dims = tuple(np.ceil(tf_data_to_img([wh_dims]))[0].astype(np.int).tolist())
+            img_wh_dims = tuple(np.ceil(tf_data_to_img([wh_dims]))[0].astype(int).tolist())
             img_dims = img_wh_dims[::-1]
         else:
             img_dims = np.array(img_dims)